requirements.txt
weaviate-client
langchain
langchain-community